Add files via upload
This is a terraform code that would provision an environment with a static page that reads: "Automation for the people" To run it, launch and ami instance and ssh into it. Install all packages as well as terraform. As an ec2-user vi into this file teera.tf to see the script of the code. Run terraform init, terraform plan and finally terraform apply. Wait for the VPC and aws ami instance to be built. SSH into the new instance and make sure that httpd is up and running. copy the public IP of the new instance and place it in your browser and hit enter. Find your static page displayed with your message thus: "Automation for the People"